// src/types/RouteName.ts
type RouteName =
  | 'adjacent-regions'
  | 'checklist-feed-on-date'
  | 'hotspot-info'
  | 'nearby-hotspots'
  | 'recent-checklists'
  | 'recent-notable-observations-region'
  | 'recent-observations-of-species-in-region'
  | 'recent-observations-region'
  | 'region-hotspots'
  | 'region-info'
  | 'region-stats-on-date'
  | 'species-list-region'
  | 'sub-region-list'
  | 'taxa-locale-codes'
  | 'taxonomic-forms'
  | 'taxonomic-groups'
  | 'taxonomy-versions'
  | 'top-100';

export default RouteName;